refactor: merge node and image create funcs into manager files
